+++ fxmanifest.lua
@@ -5,7 +5,7 @@ name 'qbxsql'
 author 'ChatDisabled'
 description 'Database adapter, compatibility layer, and schema manager for FiveM'
 version '2.14.1'
-qbxsql_version '0.3.0'
+qbxsql_version '0.3.1'
 license 'MIT'
 
 node_version '22'